refactor(src): adopt src/ layout & package `mlops_fraud`
- Move importable code under src/mlops_fraud/*
- Migrate common/{features,schemas}.py -> src/mlops_fraud/
- Create package __init__.py files for submodules
- Add pyproject.toml (editable install)
- Add streamlit/simulate_demo.py (portfolio demo)
- Remove legacy common/ and streamlit_app/
- Update tests to import mlops_fraud.*

NOTE: ensure src/mlops_fraud/monitoring/__init__.py is an empty module (no Streamlit code).
