+++ pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "mlops-fraud"
+version = "0.1.0"
+requires-python = ">=3.9"
+description = "End-to-end fraud MLOps project"
+
+[build-system]
+requires = ["setuptools>=64", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
