+++ src/mlops_fraud/__init__.py
@@ -1,0 +1,2 @@
+__all__ = []
+__version__ = '0.1.0'
